+++ Api/SocialAdapter.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace SM\Theme\Api;
+namespace Unet\SocialShare\Api;
 
 /**
  * Class SocialAdapter
- * @package SM\Theme\Api
+ * @package Unet\SocialShare\Api
  */
 interface SocialAdapter
 {
-    const FACEBOOK_SHARE_LINK = 'https://www.facebook.com/sharer/sharer.php?u=';
-    const TWITTER_SHARE_LINK = 'https://twitter.com/home?status=';
-    const GOOGLE_PLUS_SHARE_LINK = 'https://plus.google.com/share?url=';
-    const PINTEREST_SHARE_LINK = 'https://pinterest.com/pin/create/button/?url=';
+    const FACEBOOK_SHARE_LINK = 'https://www.facebook.com/sharer/sharer.php';
+    const TWITTER_SHARE_LINK = 'https://twitter.com/intent/tweet';
+    const GOOGLE_PLUS_SHARE_LINK = 'https://plus.google.com/share';
+    const PINTEREST_SHARE_LINK = 'https://pinterest.com/pin/create/button';
+
+    public function getShareConfig();
 }