+++ Block/ShareTo.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Unet\SocialShare\Block;
+
+/**
+ * Class ShareTo
+ * @package Unet\SocialShare\Block
+ */
+class ShareTo extends \Unet\SocialShare\Block\AbstractShare
+{
+
+}