+++ registration.php
@@ -2,6 +2,6 @@
 
 \Magento\Framework\Component\ComponentRegistrar::register(
 \Magento\Framework\Component\ComponentRegistrar::MODULE,
-'Social_Share',
+'Unet_SocialShare',
 __DIR__
 );