+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## 1.0.0
 
 Initial beta release
+
+- Open graph meta
+
+- Twitter card meta
+
+- Social javascript API
+
+- Share toolbar